stern_school/requirements.txt
Django==
django-tinymce==3.0.2